implemented textured 3d model loading and rendering

+++ GP3/src/time.h
@@ -2,6 +2,6 @@
 
 struct Time
 {
-	double currentMs;
-	double deltaMs;
+	float currentMs;
+	float deltaMs;
 };
